+++ include/smal/Memory/util.hpp
@@ -9,7 +9,7 @@ namespace ma
      *
      */
     char*
-    memory_set(void* memory, usize size, char byte);
+    memory_set(void* memory, u32 size, char byte);
 } // namespace ma
 
 #endif // SMAL_MEMORY_HELPER_HPP
